User IDs shouldn't be saved in a small integer field

+++ workbench/accounts/models.py
@@ -356,7 +356,7 @@ class Team(Model):
 
 class CoffeePairings(models.Model):
     created_at = models.DateTimeField(_("created at"), default=timezone.now)
-    users = ArrayField(models.SmallIntegerField(), verbose_name=_("users"))
+    users = ArrayField(models.IntegerField(), verbose_name=_("users"))
 
     class Meta:
         verbose_name = _("coffee pairings")
